[DeadCode] Add TagRemovalGuard for re-use code check that @param/@return allowed to be removed
