+++ rules/DeadCode/PhpDoc/DeadParamTagValueNodeAnalyzer.php
@@ -4,23 +4,12 @@ declare(strict_types=1);
 
 namespace Rector\DeadCode\PhpDoc;
 
-use PhpParser\Node;
 use PhpParser\Node\FunctionLike;
 use PhpParser\Node\Name;
-use PhpParser\Node\Name\FullyQualified;
 use PhpParser\Node\Param;
 use PHPStan\PhpDocParser\Ast\PhpDoc\ParamTagValueNode;
-use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocChildNode;
-use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocNode;
-use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocTagNode;
-use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocTextNode;
 use PHPStan\PhpDocParser\Ast\Type\IdentifierTypeNode;
-use Rector\BetterPhpDocParser\PhpDocManipulator\PhpDocTypeChanger;
-use Rector\BetterPhpDocParser\ValueObject\PhpDoc\VariadicAwareParamTagValueNode;
-use Rector\BetterPhpDocParser\ValueObject\PhpDocAttributeKey;
-use Rector\BetterPhpDocParser\ValueObject\Type\BracketsAwareUnionTypeNode;
-use Rector\DeadCode\TypeNodeAnalyzer\GenericTypeNodeAnalyzer;
-use Rector\DeadCode\TypeNodeAnalyzer\MixedArrayTypeNodeAnalyzer;
+use Rector\DeadCode\PhpDoc\Guard\TagRemovalGuard;
 use Rector\NodeNameResolver\NodeNameResolver;
 use Rector\NodeTypeResolver\TypeComparator\TypeComparator;
 
@@ -29,8 +18,7 @@ final class DeadParamTagValueNodeAnalyzer
     public function __construct(
         private readonly NodeNameResolver $nodeNameResolver,
         private readonly TypeComparator $typeComparator,
-        private readonly GenericTypeNodeAnalyzer $genericTypeNodeAnalyzer,
-        private readonly MixedArrayTypeNodeAnalyzer $mixedArrayTypeNodeAnalyzer
+        private readonly TagRemovalGuard $tagRemovalGuard
     ) {
     }
 
@@ -57,87 +45,7 @@ final class DeadParamTagValueNodeAnalyzer
             return false;
         }
 
-        if (in_array($paramTagValueNode->type::class, PhpDocTypeChanger::ALLOWED_TYPES, true)) {
-            return false;
-        }
-
-        if (! $paramTagValueNode->type instanceof BracketsAwareUnionTypeNode) {
-            return $this->isEmptyDescription($paramTagValueNode, $param->type);
-        }
-
-        if ($this->mixedArrayTypeNodeAnalyzer->hasMixedArrayType($paramTagValueNode->type)) {
-            return false;
-        }
-
-        if (! $this->genericTypeNodeAnalyzer->hasGenericType($paramTagValueNode->type)) {
-            return $this->isEmptyDescription($paramTagValueNode, $param->type);
-        }
-
-        return false;
-    }
-
-    private function isEmptyDescription(ParamTagValueNode $paramTagValueNode, Node $node): bool
-    {
-        if ($paramTagValueNode->description !== '') {
-            return false;
-        }
-
-        $parent = $paramTagValueNode->getAttribute(PhpDocAttributeKey::PARENT);
-        if (! $parent instanceof PhpDocTagNode) {
-            return true;
-        }
-
-        $parent = $parent->getAttribute(PhpDocAttributeKey::PARENT);
-        if (! $parent instanceof PhpDocNode) {
-            return true;
-        }
-
-        $children = $parent->children;
-
-        foreach ($children as $key => $child) {
-            if ($child instanceof PhpDocTagNode && $node instanceof FullyQualified) {
-                return $this->isUnionIdentifier($child);
-            }
-
-            if (! $this->isTextNextline($key, $child)) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    private function isTextNextline(int $key, PhpDocChildNode $phpDocChildNode): bool
-    {
-        if ($key < 1) {
-            return true;
-        }
-
-        if (! $phpDocChildNode instanceof PhpDocTextNode) {
-            return true;
-        }
-
-        return (string) $phpDocChildNode === '';
-    }
-
-    private function isUnionIdentifier(PhpDocTagNode $phpDocTagNode): bool
-    {
-        if (! $phpDocTagNode->value instanceof VariadicAwareParamTagValueNode) {
-            return true;
-        }
-
-        if (! $phpDocTagNode->value->type instanceof BracketsAwareUnionTypeNode) {
-            return true;
-        }
-
-        $types = $phpDocTagNode->value->type->types;
-        foreach ($types as $type) {
-            if ($type instanceof IdentifierTypeNode) {
-                return false;
-            }
-        }
-
-        return true;
+        return $this->tagRemovalGuard->isLegal($paramTagValueNode, $param->type);
     }
 
     private function matchParamByName(string $desiredParamName, FunctionLike $functionLike): ?Param
