+++ rules/DeadCode/PhpDoc/DeadReturnTagValueNodeAnalyzer.php
@@ -8,13 +8,9 @@ use PhpParser\Node\FunctionLike;
 use PhpParser\Node\Stmt\ClassLike;
 use PhpParser\Node\Stmt\Trait_;
 use PHPStan\PhpDocParser\Ast\PhpDoc\ReturnTagValueNode;
-use PHPStan\PhpDocParser\Ast\Type\IdentifierTypeNode;
 use PHPStan\PhpDocParser\Ast\Type\ThisTypeNode;
-use Rector\BetterPhpDocParser\PhpDocManipulator\PhpDocTypeChanger;
-use Rector\BetterPhpDocParser\ValueObject\Type\BracketsAwareUnionTypeNode;
 use Rector\Core\PhpParser\Node\BetterNodeFinder;
-use Rector\DeadCode\TypeNodeAnalyzer\GenericTypeNodeAnalyzer;
-use Rector\DeadCode\TypeNodeAnalyzer\MixedArrayTypeNodeAnalyzer;
+use Rector\DeadCode\PhpDoc\Guard\TagRemovalGuard;
 use Rector\NodeTypeResolver\TypeComparator\TypeComparator;
 
 final class DeadReturnTagValueNodeAnalyzer
@@ -22,8 +18,7 @@ final class DeadReturnTagValueNodeAnalyzer
     public function __construct(
         private readonly TypeComparator $typeComparator,
         private readonly BetterNodeFinder $betterNodeFinder,
-        private readonly GenericTypeNodeAnalyzer $genericTypeNodeAnalyzer,
-        private readonly MixedArrayTypeNodeAnalyzer $mixedArrayTypeNodeAnalyzer,
+        private readonly TagRemovalGuard $tagRemovalGuard
     ) {
     }
 
@@ -47,44 +42,6 @@ final class DeadReturnTagValueNodeAnalyzer
             return false;
         }
 
-        if (in_array($returnTagValueNode->type::class, PhpDocTypeChanger::ALLOWED_TYPES, true)) {
-            return false;
-        }
-
-        if (! $returnTagValueNode->type instanceof BracketsAwareUnionTypeNode) {
-            return $returnTagValueNode->description === '';
-        }
-
-        if ($this->genericTypeNodeAnalyzer->hasGenericType($returnTagValueNode->type)) {
-            return false;
-        }
-
-        if ($this->mixedArrayTypeNodeAnalyzer->hasMixedArrayType($returnTagValueNode->type)) {
-            return false;
-        }
-
-        if ($this->hasTruePseudoType($returnTagValueNode->type)) {
-            return false;
-        }
-
-        return $returnTagValueNode->description === '';
-    }
-
-    private function hasTruePseudoType(BracketsAwareUnionTypeNode $bracketsAwareUnionTypeNode): bool
-    {
-        $unionTypes = $bracketsAwareUnionTypeNode->types;
-
-        foreach ($unionTypes as $unionType) {
-            if (! $unionType instanceof IdentifierTypeNode) {
-                continue;
-            }
-
-            $name = strtolower((string) $unionType);
-            if ($name === 'true') {
-                return true;
-            }
-        }
-
-        return false;
+        return $this->tagRemovalGuard->isLegal($returnTagValueNode, $returnType);
     }
 }
